agentic-ai-dataset-maker/requirements.txt
streamlit
openai
pandas
pygithub